refactor serializers, views, and urls
