+++ course_discovery/urls.py
@@ -50,11 +50,6 @@ urlpatterns = auth_urlpatterns + [
     url(r'^taggit_autosuggest/', include('taggit_autosuggest.urls')),
 ]
 
-if 'course_discovery.apps.edx_catalog_extensions' in settings.INSTALLED_APPS:
-    urlpatterns.append(
-        url(r'^api-edx/', include('course_discovery.apps.edx_catalog_extensions.urls'))
-    )
-
 if settings.DEBUG:  # pragma: no cover
     # We need this url pattern to serve user uploaded assets according to
     # https://docs.djangoproject.com/en/1.10/howto/static-files/#serving-files-uploaded-by-a-user-during-development
